// src/nebular-react-styles/src/theme/types/nebular-theme/NebularThemeMenu.ts
export interface NebularThemeMenu {
  menu_background_color: any;
  menu_text_color: any;
  menu_text_font_family: any;
  menu_text_font_size: any;
  menu_text_font_weight: any;
  menu_text_line_height: any;

  menu_group_text_color: any;

  menu_item_border_radius: any;
  menu_item_padding: any;

  menu_item_hover_background_color: any;
  menu_item_hover_cursor: any;
  menu_item_hover_text_color: any;
  menu_item_icon_hover_color: any;

  menu_item_active_background_color: any;
  menu_item_active_text_color: any;
  menu_item_icon_active_color: any;

  menu_item_icon_color: any;
  menu_item_icon_margin: any;
  menu_item_icon_width: any;

  menu_item_divider_color: any;
  menu_item_divider_style: any;
  menu_item_divider_width: any;

  menu_submenu_background_color: any;
  menu_submenu_text_color: any;
  menu_submenu_margin: any;
  menu_submenu_padding: any;

  menu_submenu_item_border_color: any;
  menu_submenu_item_border_style: any;
  menu_submenu_item_border_width: any;
  menu_submenu_item_border_radius: any;
  menu_submenu_item_padding: any;

  menu_submenu_item_hover_background_color: any;
  menu_submenu_item_hover_border_color: any;
  menu_submenu_item_hover_text_color: any;
  menu_submenu_item_icon_hover_color: any;

  menu_submenu_item_active_background_color: any;
  menu_submenu_item_active_border_color: any;
  menu_submenu_item_active_text_color: any;
  menu_submenu_item_icon_active_color: any;

  menu_submenu_item_active_hover_background_color: any;
  menu_submenu_item_active_hover_border_color: any;
  menu_submenu_item_active_hover_text_color: any;
  menu_submenu_item_icon_active_hover_color: any;
}